requirements.txt
langchain
pyPDF2
python-dotenv
streamlit
faiss-cpu
google-generativeai
faiss-cpu
altair 
tiktoken 
langchain-google-genai